+++ src/uws_client.c
@@ -722,7 +722,7 @@ static void on_underlying_io_open_complete(void* context, IO_OPEN_RESULT open_re
                         "\r\n";
                     const char* base64_nonce_chars = STRING_c_str(base64_nonce);
 
-                    upgrade_request_length = strlen(upgrade_request_format) + strlen(uws_client->resource_name)+strlen(uws_client->hostname) + strlen(base64_nonce_chars) + strlen(uws_client->protocols[0].protocol)+5;
+                    upgrade_request_length = (int)(strlen(upgrade_request_format) + strlen(uws_client->resource_name)+strlen(uws_client->hostname) + strlen(base64_nonce_chars) + strlen(uws_client->protocols[0].protocol)+5);
                     if (upgrade_request_length < 0)
                     {
                         /* Codes_SRS_UWS_CLIENT_01_408: [ If constructing of the WebSocket upgrade request fails, uws shall report that the open failed by calling the `on_ws_open_complete` callback passed to `uws_client_open_async` with `WS_OPEN_ERROR_CONSTRUCTING_UPGRADE_REQUEST`. ]*/
